Refactor model classes into a separate module
