+++ user-data-store-client-model/src/main/java/com/wultra/security/userdatastore/client/model/response/DocumentCreateResponse.java
@@ -15,38 +15,21 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.wultra.security.userdatastore.model.request;
+package com.wultra.security.userdatastore.client.model.response;
 
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.Size;
 import lombok.Builder;
 import lombok.extern.jackson.Jacksonized;
 
-import java.util.Map;
-
 /**
- * Request class for updating documents.
+ * Response class for creating documents.
  *
  * @author Roman Strobl, roman.strobl@wultra.com
  */
 @Builder
 @Jacksonized
-public record DocumentUpdateRequest(
+public record DocumentCreateResponse (
 
-        @NotBlank @Size(max = 255)
-        String userId,
-        @NotBlank @Size(max = 36)
         String id,
-        @NotBlank @Size(max = 32)
-        String documentType,
-        @NotBlank @Size(max = 32)
-        String dataType,
-        @NotBlank @Size(max = 255)
-        String documentDataId,
-        @Size(max = 255)
-        String externalId,
-        @NotBlank
-        String documentData,
-        Map<String, Object> attributes
+        String documentId
 
-) { }
+) {}
