+++ user-data-store-client-model/src/main/java/com/wultra/security/userdatastore/client/model/response/PhotoResponse.java
@@ -15,21 +15,21 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.wultra.security.userdatastore.model.response;
+package com.wultra.security.userdatastore.client.model.response;
 
-import com.wultra.security.userdatastore.model.dto.DocumentDto;
+import com.wultra.security.userdatastore.client.model.dto.PhotoDto;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 
 import java.util.ArrayList;
 
 /**
- * Response class for listing documents.
+ * Response class for listing photos.
  *
  * @author Roman Strobl, roman.strobl@wultra.com
  */
 @Data
 @EqualsAndHashCode(callSuper = true)
-public class DocumentResponse extends ArrayList<DocumentDto> {
+public class PhotoResponse extends ArrayList<PhotoDto> {
 
 }
