+++ user-data-store-server/src/main/java/com/wultra/security/userdatastore/converter/PhotoConverter.java
@@ -17,8 +17,7 @@
  */
 package com.wultra.security.userdatastore.converter;
 
-import com.wultra.security.userdatastore.model.dto.AttachmentDto;
-import com.wultra.security.userdatastore.model.dto.PhotoDto;
+import com.wultra.security.userdatastore.client.model.dto.PhotoDto;
 import com.wultra.security.userdatastore.model.entity.AttachmentEntity;
 import com.wultra.security.userdatastore.model.entity.DocumentEntity;
 import com.wultra.security.userdatastore.model.entity.PhotoEntity;
