+++ user-data-store-server/src/main/java/com/wultra/security/userdatastore/service/PhotoService.java
@@ -19,16 +19,16 @@ package com.wultra.security.userdatastore.service;
 
 import com.wultra.core.audit.base.Audit;
 import com.wultra.core.audit.base.model.AuditDetail;
+import com.wultra.security.userdatastore.client.model.dto.PhotoDto;
+import com.wultra.security.userdatastore.client.model.request.PhotoCreateRequest;
+import com.wultra.security.userdatastore.client.model.response.PhotoCreateResponse;
+import com.wultra.security.userdatastore.client.model.response.PhotoResponse;
 import com.wultra.security.userdatastore.converter.PhotoConverter;
-import com.wultra.security.userdatastore.model.dto.PhotoDto;
 import com.wultra.security.userdatastore.model.entity.DocumentEntity;
 import com.wultra.security.userdatastore.model.entity.PhotoEntity;
 import com.wultra.security.userdatastore.model.error.ResourceNotFoundException;
 import com.wultra.security.userdatastore.model.repository.DocumentRepository;
 import com.wultra.security.userdatastore.model.repository.PhotoRepository;
-import com.wultra.security.userdatastore.model.request.PhotoCreateRequest;
-import com.wultra.security.userdatastore.model.response.PhotoCreateResponse;
-import com.wultra.security.userdatastore.model.response.PhotoResponse;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.core.context.SecurityContextHolder;
@@ -104,10 +104,7 @@ public class PhotoService {
 
         photoRepository.save(photoEntity);
 
-        final PhotoCreateResponse response = new PhotoCreateResponse();
-        response.setId(photoEntity.getId());
-        response.setDocumentId(documentEntity.getId());
-        return response;
+        return new PhotoCreateResponse(photoEntity.getId(), documentEntity.getId());
     }
 
     @Transactional
